+++ packages/asset-kit-block/manifest.json
