+++ packages/asset-kit-block/src/components/InformationSection.tsx
